TSF-2530 Justeringer i Fakta Beregning Periode med AP velges automatisk Formattering av tabelloverskrifter Flytter aksjonspunktmeldinger over perioder
