+++ packages/fakta-beregning/i18n/nb_NO.json
@@ -61,16 +61,16 @@
   "BeregningInfoPanel.AvklareAktiviteterField.HvordanGarJegFrem2": "Ved full AAP skal arbeidsinntekten ikke benyttes.",
 
   "BeregningInfoPanel.VurderFaktaBeregningField.ATFLSammeOrg": "<h3>Søker er arbeidstaker og frilans i samme virksomhet.</h3>Inntekter er rapportert inn på samme org. nummer, og inntektene kan ikke skilles fra hverandre. Fastsett hva som er arbeidsinntekt og hva som er samlet frilansinntekt.",
-  "BeregningInfoPanel.VurderFaktaBeregningField.VurderLonnsendringHelpText": "<h3>Vurder om søker har hatt lønnsendring</h3>Det er registrert lønnsendring 16.02.2023 for xxxx, men ny inntekt kan ikke fastsettes av systemet. Vurder om registrert lønnsendring er riktig og fastsett ny månedsinntekt.",
-  "BeregningInfoPanel.VurderFaktaBeregningField.VurderMottarYtelseHelpTextFrilans": "<h3>Vurder om søker mottar ytelse for frilansaktivitet</h3>Det er funnet utbetaling av [ytelse] i beregningsperioden, men utbetalt beløp kan ikke hentes fra registrene.",
-  "BeregningInfoPanel.VurderFaktaBeregningField.VurderMottarYtelseHelpTextArbeidstaker": "<h3>Vurder om søker mottar ytelse for arbeidsforhold</h3>Det er funnet utbetaling av [ytelse] i beregningsperioden, men utbetalt beløp kan ikke hentes fra registrene.",
+  "BeregningInfoPanel.VurderFaktaBeregningField.VurderLonnsendringHelpText": "<h3>Vurder om søker har hatt lønnsendring</h3>Det er registrert lønnsendring, men ny inntekt kan ikke fastsettes av systemet. Vurder om registrert lønnsendring er riktig og fastsett ny månedsinntekt.",
+  "BeregningInfoPanel.VurderFaktaBeregningField.VurderMottarYtelseHelpTextFrilans": "<h3>Vurder om søker mottar ytelse for frilansaktivitet</h3>Det er funnet utbetaling av ytelse i beregningsperioden, men utbetalt beløp kan ikke hentes fra registrene.",
+  "BeregningInfoPanel.VurderFaktaBeregningField.VurderMottarYtelseHelpTextArbeidstaker": "<h3>Vurder om søker mottar ytelse for arbeidsforhold</h3>Det er funnet utbetaling av ytelse i beregningsperioden, men utbetalt beløp kan ikke hentes fra registrene.",
   "BeregningInfoPanel.VurderFaktaBeregningField.VurderEtterlonnSluttpakkeHelpText": "<h3>Vurder om søker har etterlønn eller sluttvederlag i beregningsperioden</h3>Søker har oppgitt etterlønn eller sluttvederlag i søknaden.",
   "BeregningInfoPanel.VurderFaktaBeregningField.VurderMilitaerSiviltjenesteHelpText": "<h3>Vurder om søker har vært i militær- eller siviljeneste i opptjeningsperioden</h3>Søker har oppgitt militær- eller siviltjeneste i søknaden.",
-  "BeregningInfoPanel.VurderFaktaBeregningField.FastsettBGKunYtelseHelpText": "<h3>Søker har [ytelse] som eneste inntekt</h3>Det er funnet utbetaling av ytelse i beregningsperioden, men beløp og inntektskategori kan ikke hentes fra registrene.",
+  "BeregningInfoPanel.VurderFaktaBeregningField.FastsettBGKunYtelseHelpText": "<h3>Søker har ytelse som eneste inntekt</h3>Det er funnet utbetaling av ytelse i beregningsperioden, men beløp og inntektskategori kan ikke hentes fra registrene.",
   "BeregningInfoPanel.VurderFaktaBeregningField.VurderNyoppstartetFLHelpText": "<h3>Vurder om frilanser er nyoppstartet</h3>Det er funnet frilansoppdrag med oppstartsdato i beregningsperioden. Vurder om frilansoppdrag er nyoppstartet, og om det skal benyttes en kortere periode for å fastsette inntekt.",
   "BeregningInfoPanel.VurderFaktaBeregningField.VurderSNNyIArbeidslivetHelpText": "<h3>Vurder om søker ble yrkesaktiv i løpet av de siste tre årene</h3>Søker har oppgitt i søknaden at de er selvstendig næringsdrivende og ny i arbeidslivet.",
   "BeregningInfoPanel.VurderFaktaBeregningField.VurderRefusjonskravKommetForSentHelpText": "<h3>Vurder refusjonskrav som har kommet for sent</h3>{arbeidsgiverVisningsnavn} har minst ett refusjonskrav som har kommet for sent.",
-  "BeregningInfoPanel.VurderFaktaBeregningField.TidsbegrensetArbeidsforholdHelpText": "<h3>Vurder om arbeidsforholdet er tidsbegrenset</h3>Det er funnet sluttdato frem i tid for {arbeidsgiverVisningsnavn} og arbeidsforholdet er kortere enn 6 måneder.",
+  "BeregningInfoPanel.VurderFaktaBeregningField.TidsbegrensetArbeidsforholdHelpText": "<h3>Vurder om arbeidsforholdet er tidsbegrenset</h3>Det er funnet sluttdato frem i tid for arbeidsforholdet og arbeidsforholdet er kortere enn 6 måneder.",
   "BeregningInfoPanel.VurderFaktaBeregningField.FastsettBGKunYtelse.HvordanGarJegFremForFastsetteManedsinntekt1": "Gå til A-inntekt for å finne brukerens inntekter.",
   "BeregningInfoPanel.VurderFaktaBeregningField.FastsettBGKunYtelse.HvordanGarJegFremForFastsetteManedsinntekt2": "Bruk først filter §8-28 og finn samlet inntekt for de 3 siste månedene før skjæringstidspunktet (utbetalinger fra Nav ligger ikke i dette filteret, kan derfor være 0). Regn om til årsinntekt.",
   "BeregningInfoPanel.VurderFaktaBeregningField.FastsettBGKunYtelse.HvordanGarJegFremForFastsetteManedsinntekt3": "Bruk så filter §8-30 og regn ut den totale inntekten for de siste 12 månedene før skjæringstidspunktet.",
