--- packages/fakta-beregning/src/components/fellesFaktaForATFLogSN/vurderFaktaBeregningHjelpefunksjoner.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const parseStringToBoolean = (value: string) => value === 'true';
